+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 UNRELEASED
 ----------
 
+* [ [#1460](https://github.com/digitalfabrik/integreat-cms/issues/1460) ] Only show status in broken link checker for expert users
+
 
 2022.5.2
 --------