+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 UNRELEASED
 ----------
 
+* [ [#1460](https://github.com/digitalfabrik/integreat-cms/issues/1460) ] Only show status in broken link checker for expert users
+* [ [#742](https://github.com/digitalfabrik/integreat-cms/issues/742) ] Add default bounding box to region API
+* [ [#1406](https://github.com/digitalfabrik/integreat-cms/issues/1406) ] Hide sub-headings in PDF table of contents
+* [ [#1478](https://github.com/digitalfabrik/integreat-cms/issues/1478) ] Fix bug where page with archived sibling cannot be saved
+* [ [#1452](https://github.com/digitalfabrik/integreat-cms/issues/1452) ] Only allow users of the same region for page-specific-permissions
+* [ [#1481](https://github.com/digitalfabrik/integreat-cms/issues/1481) ] Support last week for monthly recurring events
+* [ [#1487](https://github.com/digitalfabrik/integreat-cms/issues/1487) ] Invalidate cache of related objects when languages are changed
+
 
 2022.5.2
 --------